2. Model Training, Evaluation & Deployment/final_retrieval_testing/requirements.txt
# Core dependencies for Retrieval Performance Testing System
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Vector search and embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# Text search and preprocessing
rank-bm25>=0.2.2
scikit-learn>=1.0.0

# Progress bars and utilities
tqdm>=4.64.0
psutil>=5.8.0

# Optional: For advanced visualizations
plotly>=5.0.0